refactor(#S11P12D203-280): ViewModel의 Area API Long값으로 리팩토링

+++ android/project/app/src/main/java/com/chill/mallang/ui/feature/map/MapViewModel.kt
@@ -85,7 +85,7 @@ class MapViewModel
             }
         }
 
-        fun loadTryCount(areaId: Int) {
+        fun loadTryCount(areaId: Long) {
             viewModelScope.launch {
                 areaRepository.getTryCount(areaId).collectLatest { response ->
                     when (response) {
